fix: issue rename file
