+++ src/components/Graphics/PickingUpBook.tsx
