+++ src/components/Graphics/SuccessBorrowing.tsx
