+++ src/components/Graphics/WelcomeToBooking.tsx
