+++ src/components/Graphics/WhyUs.tsx
