+++ src/pages/NotFound.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@components/Button'
 import { Typography } from '@components/Typograpy'
-import GraphicNotFound from '@components/Graphic/NotFound'
+import GraphicNotFound from '@components/Graphics/NotFound'
 import { ChevronLeft } from 'lucide-react'
 import type React from 'react'
 import { Link } from 'react-router-dom'
